Project code/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
attrs==23.1.0
blinker==1.6.2
blis==0.7.10
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
community==1.0.0b1
confection==0.1.1
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.7
exceptiongroup==1.1.3
fastapi==0.101.1
fastprogress==1.0.3
fasttext-wheel==0.9.2
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.42.0
fsspec==2023.6.0
gensim==4.3.2
h11==0.14.0
h5py==2.10.0
hazm==0.9.3
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imgbeddings==0.1.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
keras-resnet==0.2.0
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.3
matplotlib==3.3.2
more-itertools==10.1.0
mpmath==1.3.0
murmurhash==1.0.9
networkx==3.1
nltk==3.8.1
numpy==1.22.1
onnx==1.14.1
onnxruntime==1.15.1
onnxruntime-tools==1.7.0
opencv-python==4.8.0.76
outcome==1.2.0
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==7.0.0
pluggy==1.3.0
preshed==3.0.8
protobuf==4.24.2
psutil==5.9.5
py-cpuinfo==9.0.0
py3nvml==0.2.7
pybind11==2.11.1
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
pyparsing==3.0.9
pyreadline3==3.4.1
PySocks==1.7.1
pytest==7.4.0
python-box==7.1.1
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-louvain==0.16
python-multipart==0.0.6
pytorch-lamb==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
scikit-learn==1.3.0
scipy==1.10.1
selenium==4.11.2
sent2vec==0.3.0
SentimentAnalysis==0.8
seqeval==1.2.2
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.7
starlette==0.27.0
sympy==1.12
tensorboardX==2.6.2.2
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.12.1
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.22.2
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.23.2
vaderSentiment==3.3.2
wasabi==1.1.2
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.7
wsproto==1.2.0
xmltodict==0.13.0
zipp==3.16.2
